move functions to lib.rs from main.rs
